Make code refactory
