+++ app/src/main/java/com/msaggik/githubclientapp/di/App.kt
@@ -4,16 +4,20 @@ import android.app.Application
 import android.content.SharedPreferences
 import android.util.Log
 import androidx.appcompat.app.AppCompatDelegate
+import com.msaggik.githubclientapp.di.repository.AppComponent
+import com.msaggik.githubclientapp.di.repository.DaggerAppComponent
 
 private const val SHARED_PREFERENCES = "shared_preferences"
 private const val THEME_APP_KEY = "theme_app_key"
 class App : Application() {
 
+    lateinit var appComponent: AppComponent
     private var lightTheme = true
     private lateinit var sharedPreferences: SharedPreferences
 
     override fun onCreate() {
         super.onCreate()
+        appComponent = DaggerAppComponent.create()
         sharedPreferences = getSharedPreferences(SHARED_PREFERENCES, MODE_PRIVATE)
         lightTheme = getSharedPreferences(SHARED_PREFERENCES, MODE_PRIVATE).getBoolean(THEME_APP_KEY, false)
         setApplicationTheme(lightTheme)
