+++ app/src/main/java/com/msaggik/githubclientapp/di/repository/AppComponent.kt
@@ -1,0 +1,16 @@
+package com.msaggik.githubclientapp.di.repository
+
+import com.msaggik.githubclientapp.view.fragments.ItemFragment
+import com.msaggik.githubclientapp.view.fragments.SearchFragment
+import dagger.Component
+import javax.inject.Singleton
+
+@Component(
+    modules = [
+    ]
+)
+@Singleton
+interface AppComponent {
+//    fun injectItemFragmentRestApi(itemFragment: ItemFragment)
+//    fun injectSearchFragmentRestApi(searchFragment: SearchFragment)
+}
