+++ app/src/main/java/com/msaggik/githubclientapp/model/entities/item/User.kt
@@ -1,8 +1,8 @@
-package com.msaggik.githubclientapp.util.entities
+package com.msaggik.githubclientapp.model.entities.item
 
 import com.google.gson.annotations.SerializedName
 
-data class Follower(
+data class User(
     @SerializedName("login") var login: String? = null,
     @SerializedName("id") var id: Int? = null,
     @SerializedName("node_id") var nodeId: String? = null,
@@ -20,5 +20,19 @@ data class Follower(
     @SerializedName("events_url") var eventsUrl: String? = null,
     @SerializedName("received_events_url") var receivedEventsUrl: String? = null,
     @SerializedName("type") var type: String? = null,
-    @SerializedName("site_admin") var siteAdmin: Boolean? = null
+    @SerializedName("site_admin") var siteAdmin: Boolean? = null,
+    @SerializedName("name") var name: String? = null,
+    @SerializedName("company") var company: String? = null,
+    @SerializedName("blog") var blog: String? = null,
+    @SerializedName("location") var location: String? = null,
+    @SerializedName("email") var email: String? = null,
+    @SerializedName("hireable") var hireable: String? = null,
+    @SerializedName("bio") var bio: String? = null,
+    @SerializedName("twitter_username") var twitterUsername: String? = null,
+    @SerializedName("public_repos") var publicRepos: Int? = null,
+    @SerializedName("public_gists") var publicGists: Int? = null,
+    @SerializedName("followers") var followers: Int? = null,
+    @SerializedName("following") var following: Int? = null,
+    @SerializedName("created_at") var createdAt: String? = null,
+    @SerializedName("updated_at") var updatedAt: String? = null
 )
