+++ app/src/main/java/com/msaggik/githubclientapp/util/Util.kt
@@ -1,0 +1,4 @@
+package com.msaggik.githubclientapp.util
+
+object Util {
+}
