+++ app/src/main/res/layout/list_item_repository.xml
@@ -13,7 +13,7 @@
         style="@style/RepositoryName"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        app:layout_constraintHorizontal_weight="3"
+        app:layout_constraintHorizontal_weight="4"
         android:background="@null"
         android:text="@string/repository_name"
         android:layout_marginEnd="@dimen/dimen_8"
